server/requirements.txt
# Core FastAPI and dependencies
fastapi>=0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.23
pydantic>=2.6.0
pydantic-settings
pydantic[email]
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.29.3
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.28.1
anyio>=4.0.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
werkzeug==2.3.7

# AI and ML dependencies - PINNED VERSIONS
openai==1.55.3
anthropic==0.31.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0

# CRITICAL: Pin NumPy to 1.x for FAISS compatibility
numpy==1.26.4
# CRITICAL: Pin FAISS to compatible version
faiss-cpu==1.7.4
# CRITICAL: Pin typing_extensions for Pydantic compatibility
typing-extensions>=4.10.0

# Document processing
pypdf==4.0.1
python-docx==1.1.0
sentence-transformers==2.7.0
tiktoken==0.7.0
pandas==2.0.3
openpyxl==3.1.2
pdfplumber==0.11.7
python-magic==0.4.27
qdrant-client==1.14.3

